Engineering task #3426 : Implement DAO for BulkSubmission and IndividualRequest
git-svn-id: http://mcoltrac/svn/mcol-app/sdt/trunk@17195 e0420d7e-bc29-8447-b031-85aca2f896c5

+++ dao-api/src/main/java/uk/gov/moj/sdt/dao/api/IIndividualRequestDao.java
@@ -53,7 +53,7 @@ public interface IIndividualRequestDao extends IGenericDao
      * @throws DataAccessException Hibernate exception
      */
     IIndividualRequest getIndividualRequest (final IBulkCustomer bulkCustomer, final String customerReference,
-                                             final long dataRetention) throws DataAccessException;
+                                             final int dataRetention) throws DataAccessException;
 
     /**
      * Returns the individual request object for the given Sdt Reference Id.
